[#] 🩹 fix : Long 타입으로 변경 orderInRoom

+++ src/main/java/com/sparta/eduwithme/domain/comment/dto/CommentResponseDto.java
@@ -13,7 +13,7 @@ public class CommentResponseDto {
     private final String comment;
     private final LocalDateTime createdAt;
     private final LocalDateTime updatedAt;
-    private final Integer questionOrderInRoom;
+    private final Long questionOrderInRoom;
 
     public CommentResponseDto(Comment comment) {
         this.commentId = comment.getId();
