+++ src/main/java/com/sparta/eduwithme/domain/question/QuestionService.java
@@ -33,8 +33,8 @@ public class QuestionService {
     public QuestionResponseDto createQuestion(Long roomId, QuestionRequestDto requestDto) {
         Room room = roomService.findById(roomId);
 
-        Integer maxOrder = questionRepository.findMaxOrderInRoom(roomId);
-        int newOrder = (maxOrder == null) ? 1 : maxOrder + 1;
+        Long maxOrder = questionRepository.findMaxOrderInRoom(roomId);
+        long newOrder = (maxOrder == null) ? 1 : maxOrder + 1;
 
         Answer answer = new Answer(
                 requestDto.getAnswer().getFirst(),
